Migrate Tuist to v2 + fix paths
