+++ .tuist-bin/ProjectAutomation.framework/Versions/A/Resources/Info.plist
@@ -3,7 +3,7 @@
 <plist version="1.0">
 <dict>
 	<key>BuildMachineOSBuild</key>
-	<string>20F71</string>
+	<string>20G224</string>
 	<key>CFBundleDevelopmentRegion</key>
 	<string>en</string>
 	<key>CFBundleExecutable</key>
@@ -27,19 +27,19 @@
 	<key>DTCompiler</key>
 	<string>com.apple.compilers.llvm.clang.1_0</string>
 	<key>DTPlatformBuild</key>
-	<string>12B45b</string>
+	<string>12E507</string>
 	<key>DTPlatformName</key>
 	<string>macosx</string>
 	<key>DTPlatformVersion</key>
-	<string>11.0</string>
+	<string>11.3</string>
 	<key>DTSDKBuild</key>
-	<string>20A2408</string>
+	<string>20E214</string>
 	<key>DTSDKName</key>
-	<string>macosx11.0</string>
+	<string>macosx11.3</string>
 	<key>DTXcode</key>
-	<string>1220</string>
+	<string>1251</string>
 	<key>DTXcodeBuild</key>
-	<string>12B45b</string>
+	<string>12E507</string>
 	<key>LSMinimumSystemVersion</key>
 	<string>10.15</string>
 </dict>
