+++ .tuist-version
@@ -1,0 +1,1 @@
+2.3.2
