+++ Project.swift
@@ -18,362 +18,384 @@
 
 import ProjectDescription
 
-let baseSettings: [String: SettingValue] = [
-    "CODE_SIGN_STYLE": "Automatic",
-    "DEVELOPMENT_TEAM": "864VDCS2QY",
-    "CURRENT_PROJECT_VERSION": "11",
-    "MARKETING_VERSION": "4.1.0"
-]
+public extension SettingsDictionary {
+    func marketingVersion(_ value: String) -> SettingsDictionary {
+        merging(["MARKETING_VERSION": SettingValue(stringLiteral: value)])
+    }
 
-let fileProviderSettings = baseSettings.merging(["SWIFT_OBJC_BRIDGING_HEADER": "kDriveFileProvider/Validation/kDriveFileProvider-Bridging-Header.h", "SWIFT_ACTIVE_COMPILATION_CONDITIONS": "ISEXTENSION"]) { first, _ in first }
-let debugFileProviderSettings = baseSettings.merging(["SWIFT_OBJC_BRIDGING_HEADER": "kDriveFileProvider/Validation/kDriveFileProvider-Bridging-Header.h", "SWIFT_ACTIVE_COMPILATION_CONDITIONS": "ISEXTENSION DEBUG"]) { first, _ in first }
+    func bridgingHeader(path value: Path) -> SettingsDictionary {
+        merging(["SWIFT_OBJC_BRIDGING_HEADER": SettingValue(stringLiteral: value.pathString)])
+    }
 
-let shareExtensionSettings = baseSettings.merging(["SWIFT_ACTIVE_COMPILATION_CONDITIONS": "ISEXTENSION"]) { first, _ in first }
+    func compilationConditions(_ value: String) -> SettingsDictionary {
+        merging(["SWIFT_ACTIVE_COMPILATION_CONDITIONS": SettingValue(stringLiteral: value)])
+    }
 
-let debugShareExtensionSettings = baseSettings.merging(["SWIFT_ACTIVE_COMPILATION_CONDITIONS": "ISEXTENSION DEBUG"]) { first, _ in first }
+    func appIcon(name value: String) -> SettingsDictionary {
+        merging(["ASSETCATALOG_COMPILER_APPICON_NAME": SettingValue(stringLiteral: value)])
+    }
+}
 
-let actionExtensionSettings = baseSettings.merging(["SWIFT_ACTIVE_COMPILATION_CONDITIONS": "ISEXTENSION", "ASSETCATALOG_COMPILER_APPICON_NAME": "ExtensionIcon"]) { first, _ in first }
+let baseSettings = SettingsDictionary()
+    .automaticCodeSigning(devTeam: "864VDCS2QY")
+    .currentProjectVersion("11")
+    .marketingVersion("4.1.0")
 
-let debugActionExtensionSettings = baseSettings.merging(["SWIFT_ACTIVE_COMPILATION_CONDITIONS": "ISEXTENSION DEBUG", "ASSETCATALOG_COMPILER_APPICON_NAME": "ExtensionIcon"]) { first, _ in first }
+let fileProviderSettings = baseSettings
+    .bridgingHeader(path: "$(SRCROOT)/kDriveFileProvider/Validation/kDriveFileProvider-Bridging-Header.h")
+    .compilationConditions("ISEXTENSION")
+let debugFileProviderSettings = fileProviderSettings
+    .compilationConditions("ISEXTENSION DEBUG")
+
+let shareExtensionSettings = baseSettings
+    .compilationConditions("ISEXTENSION")
+let debugShareExtensionSettings = shareExtensionSettings
+    .compilationConditions("ISEXTENSION DEBUG")
+
+let actionExtensionSettings = baseSettings
+    .compilationConditions("ISEXTENSION")
+    .appIcon(name: "ExtensionIcon")
+let debugActionExtensionSettings = actionExtensionSettings
+    .compilationConditions("ISEXTENSION DEBUG")
 
 let project = Project(name: "kDrive",
-    packages: [
-            .package(url: "https://github.com/Alamofire/Alamofire.git", .upToNextMajor(from: "5.2.2")),
-            .package(url: "https://github.com/Infomaniak/ios-core.git", .upToNextMajor(from: "1.1.2")),
-            .package(url: "https://github.com/Infomaniak/ios-login.git", .upToNextMajor(from: "1.4.0")),
-            .package(url: "https://github.com/realm/realm-cocoa", .upToNextMajor(from: "10.0.0")),
-            .package(url: "https://github.com/SCENEE/FloatingPanel", .upToNextMajor(from: "2.0.0")),
-            .package(url: "https://github.com/onevcat/Kingfisher", .upToNextMajor(from: "7.0.0-beta.4")),
-            .package(url: "https://github.com/flowbe/MaterialOutlinedTextField", .upToNextMajor(from: "0.1.0")),
-            .package(url: "https://github.com/ProxymanApp/atlantis", .upToNextMajor(from: "1.3.0")),
-            .package(url: "https://github.com/ra1028/DifferenceKit", .upToNextMajor(from: "1.1.5")),
-            .package(url: "https://github.com/immortal79/LocalizeKit", .upToNextMajor(from: "1.0.1")),
-            .package(url: "https://github.com/airbnb/lottie-ios.git", .upToNextMajor(from: "3.1.9")),
-            .package(url: "https://github.com/CocoaLumberjack/CocoaLumberjack.git", .upToNextMajor(from: "3.7.0")),
-            .package(url: "https://github.com/RomanTysiachnik/DropDown.git", .branch("master")),
-            .package(url: "https://github.com/PhilippeWeidmann/SnackBar.swift", .upToNextMajor(from: "0.1.2")),
-            .package(url: "https://github.com/flowbe/SwiftRegex.git", .upToNextMajor(from: "1.0.0")),
-            .package(url: "https://github.com/getsentry/sentry-cocoa", .upToNextMajor(from: "7.2.9")),
-            .package(url: "https://github.com/adam-fowler/mqtt-nio", .upToNextMajor(from: "2.4.0")),
-            .package(url: "https://github.com/firebase/firebase-ios-sdk.git", .upToNextMajor(from: "8.0.0")),
-            .package(url: "https://github.com/airbnb/HorizonCalendar.git", .upToNextMajor(from: "1.0.0")),
-            .package(url: "https://github.com/Cocoanetics/Kvitto", .upToNextMajor(from: "1.0.0"))
-    ],
-    targets: [
-        Target(name: "kDrive",
-            platform: .iOS,
-            product: .app,
-            bundleId: "com.infomaniak.drive",
-            deploymentTarget: .iOS(targetVersion: "13.0", devices: [.iphone, .ipad]),
-            infoPlist: "kDrive/Resources/Info.plist",
-            sources: "kDrive/**",
-            resources: [
-                "kDrive/**/*.storyboard",
-                "kDrive/**/*.xcassets",
-                "kDrive/**/*.strings",
-                "kDrive/**/*.stringsdict",
-                "kDrive/**/*.xib",
-                "kDrive/**/*.json",
-                "kDrive/IAP/ProductIds.plist",
-                "kDriveCore/GoogleService-Info.plist"
-            ],
-            entitlements: "kDrive/Resources/kDrive.entitlements",
-            actions: [
-                    .post(tool: "./scripts/lint.sh", arguments: [""], name: "Swiftlint")
-            ],
-            dependencies: [
-                    .target(name: "kDriveFileProvider"),
-                    .target(name: "kDriveCore"),
-                    .target(name: "kDriveShareExtension"),
-                    .target(name: "kDriveActionExtension"),
-                    .package(product: "FloatingPanel"),
-                    .package(product: "Lottie"),
-                    .package(product: "DropDown"),
-                    .package(product: "HorizonCalendar"),
-                    .package(product: "Kvitto"),
-                    .sdk(name: "StoreKit.framework", status: .required)
-            ],
-            settings: Settings(base: baseSettings)),
-        Target(name: "kDriveTests",
-            platform: .iOS,
-            product: .unitTests,
-            bundleId: "com.infomaniak.drive.tests",
-            deploymentTarget: .iOS(targetVersion: "13.0", devices: [.iphone]),
-            infoPlist: "kDriveTests/Tests.plist",
-            sources: "kDriveTests/**",
-            dependencies: [
-                    .target(name: "kDrive")
-            ],
-            settings: Settings(base: ["CODE_SIGN_IDENTITY": "", "CODE_SIGNING_REQUIRED": "NO"])),
-        Target(name: "kDriveUITests",
-            platform: .iOS,
-            product: .uiTests,
-            bundleId: "com.infomaniak.drive.uitests",
-            deploymentTarget: .iOS(targetVersion: "13.0", devices: [.iphone]),
-            infoPlist: "kDriveTests/Tests.plist",
-            sources: "kDriveUITests/**",
-            dependencies: [
-                    .target(name: "kDrive"),
-                    .target(name: "kDriveCore")
-            ]),
-        Target(name: "kDriveCore",
-            platform: .iOS,
-            product: .framework,
-            bundleId: "com.infomaniak.drive.core",
-            deploymentTarget: .iOS(targetVersion: "13.0", devices: [.iphone, .ipad]),
-            infoPlist: "kDriveCore/Info.plist",
-            sources: "kDriveCore/**",
-            resources: [
-                "kDrive/**/*.xcassets",
-                "kDrive/**/*.strings",
-                "kDrive/**/*.stringsdict",
-                "kDriveCore/**/*.storyboard",
-                "kDriveCore/**/*.xcassets",
-                "kDriveCore/**/*.xib",
-                "kDriveCore/**/*.json",
-                "kDriveCore/GoogleService-Info.plist"
-            ],
-            dependencies: [
-                    .package(product: "Alamofire"),
-                    .package(product: "Atlantis"),
-                    .package(product: "MQTTNIO"),
-                    .package(product: "InfomaniakCore"),
-                    .package(product: "InfomaniakLogin"),
-                    .package(product: "RealmSwift"),
-                    .package(product: "LocalizeKit"),
-                    .package(product: "Kingfisher"),
-                    .package(product: "DifferenceKit"),
-                    .package(product: "CocoaLumberjackSwift"),
-                    .package(product: "MaterialOutlinedTextField"),
-                    .package(product: "SnackBar"),
-                    .package(product: "SwiftRegex"),
-                    .package(product: "Sentry"),
-                    .package(product: "FirebaseMessaging")
-            ]),
-        Target(name: "kDriveFileProvider",
-            platform: .iOS,
-            product: .appExtension,
-            bundleId: "com.infomaniak.drive.FileProvider",
-            deploymentTarget: .iOS(targetVersion: "13.0", devices: [.iphone, .ipad]),
-            infoPlist: .extendingDefault(with: [
-                "CFBundleVersion": "$(CURRENT_PROJECT_VERSION)",
-                "CFBundleShortVersionString": "$(MARKETING_VERSION)",
-                "CFBundleDisplayName": "kDrive",
-                "AppIdentifierPrefix": "$(AppIdentifierPrefix)",
-                "NSExtension": [
-                    "NSExtensionPointIdentifier": "com.apple.fileprovider-nonui",
-                    "NSExtensionPrincipalClass": "$(PRODUCT_MODULE_NAME).FileProviderExtension",
-                    "NSExtensionFileProviderDocumentGroup": "group.com.infomaniak.drive",
-                    "NSExtensionFileProviderSupportsEnumeration": true
-                ]
-                ]),
-            sources: "kDriveFileProvider/**",
-            headers: Headers(project: "kDriveFileProvider/**"),
-            entitlements: "kDriveFileProvider/FileProvider.entitlements",
-            dependencies: [
-                    .target(name: "kDriveCore")
-            ],
-            settings: Settings(base: fileProviderSettings, debug: Configuration(settings: debugFileProviderSettings))),
-        Target(name: "kDriveShareExtension",
-            platform: .iOS,
-            product: .appExtension,
-            bundleId: "com.infomaniak.drive.ShareExtension",
-            deploymentTarget: .iOS(targetVersion: "13.0", devices: [.iphone, .ipad]),
-            infoPlist: .extendingDefault(with: [
-                "CFBundleVersion": "$(CURRENT_PROJECT_VERSION)",
-                "CFBundleShortVersionString": "$(MARKETING_VERSION)",
-                "CFBundleDisplayName": "$(PRODUCT_NAME)",
-                "AppIdentifierPrefix": "$(AppIdentifierPrefix)",
-                "NSExtension": [
-                    "NSExtensionMainStoryboard": "MainInterface",
-                    "NSExtensionPointIdentifier": "com.apple.share-services",
-                    "NSExtensionAttributes": ["NSExtensionActivationRule": "SUBQUERY (extensionItems, $extensionItem, SUBQUERY ($extensionItem.attachments, $attachment, (ANY $attachment.registeredTypeIdentifiers UTI-CONFORMS-TO \"public.data\")).@count == $extensionItem.attachments.@count ).@count > 0"]
-                ]
-                ]),
-            sources: [
-                "kDriveShareExtension/**",
-                "kDrive/UI/Controller/FloatingPanelSelectOptionViewController.swift",
-                "kDrive/UI/Controller/Create File/FloatingPanelUtils.swift",
-                "kDrive/UI/Controller/Files/Categories/**",
-                "kDrive/UI/Controller/Files/Rights and Share/**",
-                "kDrive/UI/Controller/Files/Save File/**",
-                "kDrive/UI/Controller/Files/Search/**",
-                "kDrive/UI/Controller/Files/MultipleSelectionViewController.swift",
-                "kDrive/UI/Controller/Files/FileListViewController.swift",
-                "kDrive/UI/Controller/Files/FloatingPanelSortOptionTableViewController.swift",
-                "kDrive/UI/Controller/Floating Panel Information/**",
-                "kDrive/UI/Controller/NewFolder/**",
-                "kDrive/UI/Controller/Storyboard.swift",
-                "kDrive/UI/View/EmptyTableView/**",
-                "kDrive/UI/View/Header view/**",
-                "kDrive/UI/View/Generic/**",
-                "kDrive/UI/View/Files/Categories/**",
-                "kDrive/UI/View/Files/FileDetail/ShareLink/**",
-                "kDrive/UI/View/Files/SaveFile/**",
-                "kDrive/UI/View/Files/Search/**",
-                "kDrive/UI/View/Files/Upload/**",
-                "kDrive/UI/View/Files/FileCollectionViewCell.swift",
-                "kDrive/UI/View/Files/FileGridCollectionViewCell.swift",
-                "kDrive/UI/View/Files/SwipableCell.swift",
-                "kDrive/UI/View/Files/SwipableCollectionView.swift",
-                "kDrive/UI/View/Files/FloatingPanel/FloatingPanelSortOptionTableViewCell.swift",
-                "kDrive/UI/View/Files/FloatingPanel/FloatingPanelQuickActionCollectionViewCell.swift",
-                "kDrive/UI/View/Files/FloatingPanel/FloatingPanelTableViewCell.swift",
-                "kDrive/UI/View/Footer view/**",
-                "kDrive/UI/View/Menu/SwitchUser/**",
-                "kDrive/UI/View/Menu/MenuTableViewCell.swift",
-                "kDrive/UI/View/NewFolder/**",
-                "kDrive/Utils/**",
-                "Derived/Sources/Assets+KDrive.swift",
-                "Derived/Sources/Bundle+kDrive.swift",
-                "Derived/Sources/Strings+kDrive.swift"
-            ],
-            resources: [
-                "kDriveShareExtension/**/*.storyboard",
-                "kDrive/UI/Controller/Files/**/*.storyboard",
-                "kDrive/UI/Controller/Floating Panel Information/*.storyboard",
-                "kDrive/UI/Controller/NewFolder/*.storyboard",
-                "kDrive/UI/View/EmptyTableView/**/*.xib",
-                "kDrive/UI/View/Header view/**/*.xib",
-                "kDrive/UI/View/Generic/**/*.xib",
-                "kDrive/UI/View/Files/FileCollectionViewCell.xib",
-                "kDrive/UI/View/Files/FileGridCollectionViewCell.xib",
-                "kDrive/UI/View/Files/Categories/**/*.xib",
-                "kDrive/UI/View/Files/FileDetail/ShareLink/*.xib",
-                "kDrive/UI/View/Files/SaveFile/*.xib",
-                "kDrive/UI/View/Files/Search/*.xib",
-                "kDrive/UI/View/Files/Upload/*.xib",
-                "kDrive/UI/View/Files/FloatingPanel/FloatingPanelSortOptionTableViewCell.xib",
-                "kDrive/UI/View/Files/FloatingPanel/FloatingPanelQuickActionCollectionViewCell.xib",
-                "kDrive/UI/View/Files/FloatingPanel/FloatingPanelTableViewCell.xib",
-                "kDrive/UI/View/Footer view/*.xib",
-                "kDrive/UI/View/Menu/MenuTableViewCell.xib",
-                "kDrive/UI/View/Menu/SwitchUser/*.xib",
-                "kDrive/UI/View/NewFolder/*.xib",
-                "kDrive/**/*.xcassets",
-                "kDrive/**/*.strings",
-                "kDrive/**/*.stringsdict",
-                "kDrive/**/*.json"
-            ],
-            entitlements: "kDriveShareExtension/ShareExtension.entitlements",
-            actions: [
-                /* This prevents Tuist from generating automatic resources definition for this extension
-                  as disabling it seems only possible at a project level (.disableSynthesizedResourceAccessors */
-                    .pre(tool: "./scripts/fix-tuist.sh", arguments: [""], name: "Fix Tuist"),
-                    .post(tool: "./scripts/lint.sh", arguments: [""], name: "Swiftlint")
-            ],
-            dependencies: [
-                    .target(name: "kDriveCore"),
-                    .package(product: "FloatingPanel"),
-                    .package(product: "Lottie"),
-                    .package(product: "DropDown"),
-                    .package(product: "HorizonCalendar")
-            ],
-            settings: Settings(base: shareExtensionSettings, debug: Configuration(settings: debugShareExtensionSettings))),
-        Target(name: "kDriveActionExtension",
-            platform: .iOS,
-            product: .appExtension,
-            bundleId: "com.infomaniak.drive.ActionExtension",
-            deploymentTarget: .iOS(targetVersion: "13.0", devices: [.iphone, .ipad]),
-            infoPlist: .extendingDefault(with: [
-                "CFBundleVersion": "$(CURRENT_PROJECT_VERSION)",
-                "CFBundleShortVersionString": "$(MARKETING_VERSION)",
-                "CFBundleDisplayName": "Enregistrer dans kDrive",
-                "AppIdentifierPrefix": "$(AppIdentifierPrefix)",
-                "NSExtension": [
-                    "NSExtensionMainStoryboard": "MainInterface",
-                    "NSExtensionPointIdentifier": "com.apple.ui-services",
-                    "NSExtensionAttributes": [
-                        "NSExtensionActivationRule": "SUBQUERY (extensionItems, $extensionItem, SUBQUERY ($extensionItem.attachments, $attachment, (ANY $attachment.registeredTypeIdentifiers UTI-CONFORMS-TO \"public.data\")).@count == $extensionItem.attachments.@count ).@count > 0",
-                        "NSExtensionServiceAllowsFinderPreviewItem": true,
-                        "NSExtensionServiceAllowsTouchBarItem": true,
-                        "NSExtensionServiceFinderPreviewIconName": "NSActionTemplate",
-                        "NSExtensionServiceTouchBarBezelColorName": "TouchBarBezel",
-                        "NSExtensionServiceTouchBarIconName": "NSActionTemplate"
-                    ]
-                ]
-                ]),
-            sources: [
-                "kDriveActionExtension/**",
-                "kDrive/UI/Controller/FloatingPanelSelectOptionViewController.swift",
-                "kDrive/UI/Controller/Create File/FloatingPanelUtils.swift",
-                "kDrive/UI/Controller/Files/Categories/**",
-                "kDrive/UI/Controller/Files/Rights and Share/**",
-                "kDrive/UI/Controller/Files/Save File/**",
-                "kDrive/UI/Controller/Files/Search/**",
-                "kDrive/UI/Controller/Files/MultipleSelectionViewController.swift",
-                "kDrive/UI/Controller/Files/FileListViewController.swift",
-                "kDrive/UI/Controller/Files/FloatingPanelSortOptionTableViewController.swift",
-                "kDrive/UI/Controller/Floating Panel Information/**",
-                "kDrive/UI/Controller/NewFolder/**",
-                "kDrive/UI/Controller/Storyboard.swift",
-                "kDrive/UI/View/EmptyTableView/**",
-                "kDrive/UI/View/Header view/**",
-                "kDrive/UI/View/Generic/**",
-                "kDrive/UI/View/Files/Categories/**",
-                "kDrive/UI/View/Files/FileDetail/ShareLink/**",
-                "kDrive/UI/View/Files/SaveFile/**",
-                "kDrive/UI/View/Files/Search/**",
-                "kDrive/UI/View/Files/Upload/**",
-                "kDrive/UI/View/Files/FileCollectionViewCell.swift",
-                "kDrive/UI/View/Files/FileGridCollectionViewCell.swift",
-                "kDrive/UI/View/Files/SwipableCell.swift",
-                "kDrive/UI/View/Files/SwipableCollectionView.swift",
-                "kDrive/UI/View/Files/FloatingPanel/FloatingPanelSortOptionTableViewCell.swift",
-                "kDrive/UI/View/Files/FloatingPanel/FloatingPanelQuickActionCollectionViewCell.swift",
-                "kDrive/UI/View/Files/FloatingPanel/FloatingPanelTableViewCell.swift",
-                "kDrive/UI/View/Footer view/**",
-                "kDrive/UI/View/Menu/SwitchUser/**",
-                "kDrive/UI/View/Menu/MenuTableViewCell.swift",
-                "kDrive/UI/View/NewFolder/**",
-                "kDrive/Utils/**",
-                "Derived/Sources/Assets+KDrive.swift",
-                "Derived/Sources/Bundle+kDrive.swift",
-                "Derived/Sources/Strings+kDrive.swift"
-            ],
-            resources: [
-                "kDriveActionExtension/**/*.storyboard",
-                "kDrive/UI/Controller/Files/**/*.storyboard",
-                "kDrive/UI/Controller/Floating Panel Information/*.storyboard",
-                "kDrive/UI/Controller/NewFolder/*.storyboard",
-                "kDrive/UI/View/EmptyTableView/**/*.xib",
-                "kDrive/UI/View/Header view/**/*.xib",
-                "kDrive/UI/View/Generic/**/*.xib",
-                "kDrive/UI/View/Files/FileCollectionViewCell.xib",
-                "kDrive/UI/View/Files/FileGridCollectionViewCell.xib",
-                "kDrive/UI/View/Files/Categories/**/*.xib",
-                "kDrive/UI/View/Files/FileDetail/ShareLink/*.xib",
-                "kDrive/UI/View/Files/SaveFile/*.xib",
-                "kDrive/UI/View/Files/Search/*.xib",
-                "kDrive/UI/View/Files/Upload/*.xib",
-                "kDrive/UI/View/Files/FloatingPanel/FloatingPanelSortOptionTableViewCell.xib",
-                "kDrive/UI/View/Files/FloatingPanel/FloatingPanelQuickActionCollectionViewCell.xib",
-                "kDrive/UI/View/Files/FloatingPanel/FloatingPanelTableViewCell.xib",
-                "kDrive/UI/View/Footer view/*.xib",
-                "kDrive/UI/View/Menu/MenuTableViewCell.xib",
-                "kDrive/UI/View/Menu/SwitchUser/*.xib",
-                "kDrive/UI/View/NewFolder/*.xib",
-                "kDriveActionExtension/**/*.xcassets",
-                "kDrive/**/*.xcassets",
-                "kDriveActionExtension/**/*.strings",
-                "kDrive/**/Localizable.strings",
-                "kDrive/**/*.stringsdict",
-                "kDrive/**/*.json"
-            ],
-            entitlements: "kDriveActionExtension/ActionExtension.entitlements",
-            actions: [
-                /* This prevents Tuist from generating automatic resources definition for this extension
-                  as disabling it seems only possible at a project level (.disableSynthesizedResourceAccessors */
-                    .pre(tool: "./scripts/fix-tuist.sh", arguments: [""], name: "Fix Tuist"),
-                    .post(tool: "./scripts/lint.sh", arguments: [""], name: "Swiftlint")
-            ],
-            dependencies: [
-                    .target(name: "kDriveCore"),
-                    .package(product: "FloatingPanel"),
-                    .package(product: "Lottie"),
-                    .package(product: "DropDown"),
-                    .package(product: "HorizonCalendar")
-            ],
-            settings: Settings(base: actionExtensionSettings, debug: Configuration(settings: debugActionExtensionSettings)))
-    ],
-    fileHeaderTemplate: .file("file-header-template.txt"))
+                      packages: [
+                          .package(url: "https://github.com/Alamofire/Alamofire.git", .upToNextMajor(from: "5.2.2")),
+                          .package(url: "https://github.com/Infomaniak/ios-core.git", .upToNextMajor(from: "1.1.2")),
+                          .package(url: "https://github.com/Infomaniak/ios-login.git", .upToNextMajor(from: "1.4.0")),
+                          .package(url: "https://github.com/realm/realm-cocoa", .upToNextMajor(from: "10.0.0")),
+                          .package(url: "https://github.com/SCENEE/FloatingPanel", .upToNextMajor(from: "2.0.0")),
+                          .package(url: "https://github.com/onevcat/Kingfisher", .upToNextMajor(from: "7.0.0-beta.4")),
+                          .package(url: "https://github.com/flowbe/MaterialOutlinedTextField", .upToNextMajor(from: "0.1.0")),
+                          .package(url: "https://github.com/ProxymanApp/atlantis", .upToNextMajor(from: "1.3.0")),
+                          .package(url: "https://github.com/ra1028/DifferenceKit", .upToNextMajor(from: "1.1.5")),
+                          .package(url: "https://github.com/immortal79/LocalizeKit", .upToNextMajor(from: "1.0.1")),
+                          .package(url: "https://github.com/airbnb/lottie-ios.git", .upToNextMajor(from: "3.1.9")),
+                          .package(url: "https://github.com/CocoaLumberjack/CocoaLumberjack.git", .upToNextMajor(from: "3.7.0")),
+                          .package(url: "https://github.com/RomanTysiachnik/DropDown.git", .branch("master")),
+                          .package(url: "https://github.com/PhilippeWeidmann/SnackBar.swift", .upToNextMajor(from: "0.1.2")),
+                          .package(url: "https://github.com/flowbe/SwiftRegex.git", .upToNextMajor(from: "1.0.0")),
+                          .package(url: "https://github.com/getsentry/sentry-cocoa", .upToNextMajor(from: "7.2.9")),
+                          .package(url: "https://github.com/adam-fowler/mqtt-nio", .upToNextMajor(from: "2.4.0")),
+                          .package(url: "https://github.com/firebase/firebase-ios-sdk.git", .upToNextMajor(from: "8.0.0")),
+                          .package(url: "https://github.com/airbnb/HorizonCalendar.git", .upToNextMajor(from: "1.0.0")),
+                          .package(url: "https://github.com/Cocoanetics/Kvitto", .upToNextMajor(from: "1.0.0"))
+                      ],
+                      targets: [
+                          Target(name: "kDrive",
+                                 platform: .iOS,
+                                 product: .app,
+                                 bundleId: "com.infomaniak.drive",
+                                 deploymentTarget: .iOS(targetVersion: "13.0", devices: [.iphone, .ipad]),
+                                 infoPlist: .file(path: "kDrive/Resources/Info.plist"),
+                                 sources: "kDrive/**",
+                                 resources: [
+                                     "kDrive/**/*.storyboard",
+                                     "kDrive/**/*.xcassets",
+                                     "kDrive/**/*.strings",
+                                     "kDrive/**/*.stringsdict",
+                                     "kDrive/**/*.xib",
+                                     "kDrive/**/*.json",
+                                     "kDrive/IAP/ProductIds.plist",
+                                     "kDriveCore/GoogleService-Info.plist"
+                                 ],
+                                 entitlements: "kDrive/Resources/kDrive.entitlements",
+                                 scripts: [
+                                     .post(path: "scripts/lint.sh", name: "Swiftlint")
+                                 ],
+                                 dependencies: [
+                                     .target(name: "kDriveFileProvider"),
+                                     .target(name: "kDriveCore"),
+                                     .target(name: "kDriveShareExtension"),
+                                     .target(name: "kDriveActionExtension"),
+                                     .package(product: "FloatingPanel"),
+                                     .package(product: "Lottie"),
+                                     .package(product: "DropDown"),
+                                     .package(product: "HorizonCalendar"),
+                                     .package(product: "Kvitto"),
+                                     .sdk(name: "StoreKit.framework", status: .required)
+                                 ],
+                                 settings: .settings(base: baseSettings)),
+                          Target(name: "kDriveTests",
+                                 platform: .iOS,
+                                 product: .unitTests,
+                                 bundleId: "com.infomaniak.drive.tests",
+                                 deploymentTarget: .iOS(targetVersion: "13.0", devices: [.iphone]),
+                                 infoPlist: .file(path: "kDriveTests/Tests.plist"),
+                                 sources: "kDriveTests/**",
+                                 dependencies: [
+                                     .target(name: "kDrive")
+                                 ],
+                                 settings: .settings(base: baseSettings)),
+                          Target(name: "kDriveUITests",
+                                 platform: .iOS,
+                                 product: .uiTests,
+                                 bundleId: "com.infomaniak.drive.uitests",
+                                 deploymentTarget: .iOS(targetVersion: "13.0", devices: [.iphone]),
+                                 infoPlist: .file(path: "kDriveTests/Tests.plist"),
+                                 sources: "kDriveUITests/**",
+                                 dependencies: [
+                                     .target(name: "kDrive"),
+                                     .target(name: "kDriveCore")
+                                 ]),
+                          Target(name: "kDriveCore",
+                                 platform: .iOS,
+                                 product: .framework,
+                                 bundleId: "com.infomaniak.drive.core",
+                                 deploymentTarget: .iOS(targetVersion: "13.0", devices: [.iphone, .ipad]),
+                                 infoPlist: .file(path: "kDriveCore/Info.plist"),
+                                 sources: "kDriveCore/**",
+                                 resources: [
+                                     "kDrive/**/*.xcassets",
+                                     "kDrive/**/*.strings",
+                                     "kDrive/**/*.stringsdict",
+                                     "kDriveCore/**/*.storyboard",
+                                     "kDriveCore/**/*.xcassets",
+                                     "kDriveCore/**/*.xib",
+                                     "kDriveCore/**/*.json",
+                                     "kDriveCore/GoogleService-Info.plist"
+                                 ],
+                                 dependencies: [
+                                     .package(product: "Alamofire"),
+                                     .package(product: "Atlantis"),
+                                     .package(product: "MQTTNIO"),
+                                     .package(product: "InfomaniakCore"),
+                                     .package(product: "InfomaniakLogin"),
+                                     .package(product: "RealmSwift"),
+                                     .package(product: "LocalizeKit"),
+                                     .package(product: "Kingfisher"),
+                                     .package(product: "DifferenceKit"),
+                                     .package(product: "CocoaLumberjackSwift"),
+                                     .package(product: "MaterialOutlinedTextField"),
+                                     .package(product: "SnackBar"),
+                                     .package(product: "SwiftRegex"),
+                                     .package(product: "Sentry"),
+                                     .package(product: "FirebaseMessaging")
+                                 ]),
+                          Target(name: "kDriveFileProvider",
+                                 platform: .iOS,
+                                 product: .appExtension,
+                                 bundleId: "com.infomaniak.drive.FileProvider",
+                                 deploymentTarget: .iOS(targetVersion: "13.0", devices: [.iphone, .ipad]),
+                                 infoPlist: .extendingDefault(with: [
+                                     "CFBundleVersion": "$(CURRENT_PROJECT_VERSION)",
+                                     "CFBundleShortVersionString": "$(MARKETING_VERSION)",
+                                     "CFBundleDisplayName": "kDrive",
+                                     "AppIdentifierPrefix": "$(AppIdentifierPrefix)",
+                                     "NSExtension": [
+                                         "NSExtensionPointIdentifier": "com.apple.fileprovider-nonui",
+                                         "NSExtensionPrincipalClass": "$(PRODUCT_MODULE_NAME).FileProviderExtension",
+                                         "NSExtensionFileProviderDocumentGroup": "group.com.infomaniak.drive",
+                                         "NSExtensionFileProviderSupportsEnumeration": true
+                                     ]
+                                 ]),
+                                 sources: "kDriveFileProvider/**",
+                                 headers: Headers(project: "kDriveFileProvider/**"),
+                                 entitlements: "kDriveFileProvider/FileProvider.entitlements",
+                                 dependencies: [
+                                     .target(name: "kDriveCore")
+                                 ],
+                                 settings: .settings(base: fileProviderSettings, debug: debugFileProviderSettings)),
+                          Target(name: "kDriveShareExtension",
+                                 platform: .iOS,
+                                 product: .appExtension,
+                                 bundleId: "com.infomaniak.drive.ShareExtension",
+                                 deploymentTarget: .iOS(targetVersion: "13.0", devices: [.iphone, .ipad]),
+                                 infoPlist: .extendingDefault(with: [
+                                     "CFBundleVersion": "$(CURRENT_PROJECT_VERSION)",
+                                     "CFBundleShortVersionString": "$(MARKETING_VERSION)",
+                                     "CFBundleDisplayName": "$(PRODUCT_NAME)",
+                                     "AppIdentifierPrefix": "$(AppIdentifierPrefix)",
+                                     "NSExtension": [
+                                         "NSExtensionMainStoryboard": "MainInterface",
+                                         "NSExtensionPointIdentifier": "com.apple.share-services",
+                                         "NSExtensionAttributes": ["NSExtensionActivationRule": "SUBQUERY (extensionItems, $extensionItem, SUBQUERY ($extensionItem.attachments, $attachment, (ANY $attachment.registeredTypeIdentifiers UTI-CONFORMS-TO \"public.data\")).@count == $extensionItem.attachments.@count ).@count > 0"]
+                                     ]
+                                 ]),
+                                 sources: [
+                                     "kDriveShareExtension/**",
+                                     "kDrive/UI/Controller/FloatingPanelSelectOptionViewController.swift",
+                                     "kDrive/UI/Controller/Create File/FloatingPanelUtils.swift",
+                                     "kDrive/UI/Controller/Files/Categories/**",
+                                     "kDrive/UI/Controller/Files/Rights and Share/**",
+                                     "kDrive/UI/Controller/Files/Save File/**",
+                                     "kDrive/UI/Controller/Files/Search/**",
+                                     "kDrive/UI/Controller/Files/MultipleSelectionViewController.swift",
+                                     "kDrive/UI/Controller/Files/FileListViewController.swift",
+                                     "kDrive/UI/Controller/Files/FloatingPanelSortOptionTableViewController.swift",
+                                     "kDrive/UI/Controller/Floating Panel Information/**",
+                                     "kDrive/UI/Controller/NewFolder/**",
+                                     "kDrive/UI/Controller/Storyboard.swift",
+                                     "kDrive/UI/View/EmptyTableView/**",
+                                     "kDrive/UI/View/Header view/**",
+                                     "kDrive/UI/View/Generic/**",
+                                     "kDrive/UI/View/Files/Categories/**",
+                                     "kDrive/UI/View/Files/FileDetail/ShareLink/**",
+                                     "kDrive/UI/View/Files/SaveFile/**",
+                                     "kDrive/UI/View/Files/Search/**",
+                                     "kDrive/UI/View/Files/Upload/**",
+                                     "kDrive/UI/View/Files/FileCollectionViewCell.swift",
+                                     "kDrive/UI/View/Files/FileGridCollectionViewCell.swift",
+                                     "kDrive/UI/View/Files/SwipableCell.swift",
+                                     "kDrive/UI/View/Files/SwipableCollectionView.swift",
+                                     "kDrive/UI/View/Files/FloatingPanel/FloatingPanelSortOptionTableViewCell.swift",
+                                     "kDrive/UI/View/Files/FloatingPanel/FloatingPanelQuickActionCollectionViewCell.swift",
+                                     "kDrive/UI/View/Files/FloatingPanel/FloatingPanelTableViewCell.swift",
+                                     "kDrive/UI/View/Footer view/**",
+                                     "kDrive/UI/View/Menu/SwitchUser/**",
+                                     "kDrive/UI/View/Menu/MenuTableViewCell.swift",
+                                     "kDrive/UI/View/NewFolder/**",
+                                     "kDrive/Utils/**",
+                                     "Derived/Sources/Assets+KDrive.swift",
+                                     "Derived/Sources/Bundle+kDrive.swift",
+                                     "Derived/Sources/Strings+kDrive.swift"
+                                 ],
+                                 resources: [
+                                     "kDriveShareExtension/**/*.storyboard",
+                                     "kDrive/UI/Controller/Files/**/*.storyboard",
+                                     "kDrive/UI/Controller/Floating Panel Information/*.storyboard",
+                                     "kDrive/UI/Controller/NewFolder/*.storyboard",
+                                     "kDrive/UI/View/EmptyTableView/**/*.xib",
+                                     "kDrive/UI/View/Header view/**/*.xib",
+                                     "kDrive/UI/View/Generic/**/*.xib",
+                                     "kDrive/UI/View/Files/FileCollectionViewCell.xib",
+                                     "kDrive/UI/View/Files/FileGridCollectionViewCell.xib",
+                                     "kDrive/UI/View/Files/Categories/**/*.xib",
+                                     "kDrive/UI/View/Files/FileDetail/ShareLink/*.xib",
+                                     "kDrive/UI/View/Files/SaveFile/*.xib",
+                                     "kDrive/UI/View/Files/Search/*.xib",
+                                     "kDrive/UI/View/Files/Upload/*.xib",
+                                     "kDrive/UI/View/Files/FloatingPanel/FloatingPanelSortOptionTableViewCell.xib",
+                                     "kDrive/UI/View/Files/FloatingPanel/FloatingPanelQuickActionCollectionViewCell.xib",
+                                     "kDrive/UI/View/Files/FloatingPanel/FloatingPanelTableViewCell.xib",
+                                     "kDrive/UI/View/Footer view/*.xib",
+                                     "kDrive/UI/View/Menu/MenuTableViewCell.xib",
+                                     "kDrive/UI/View/Menu/SwitchUser/*.xib",
+                                     "kDrive/UI/View/NewFolder/*.xib",
+                                     "kDrive/**/*.xcassets",
+                                     "kDrive/**/*.strings",
+                                     "kDrive/**/*.stringsdict",
+                                     "kDrive/**/*.json"
+                                 ],
+                                 entitlements: "kDriveShareExtension/ShareExtension.entitlements",
+                                 scripts: [
+                                     /* This prevents Tuist from generating automatic resources definition for this extension
+                                      as disabling it seems only possible at a project level (.disableSynthesizedResourceAccessors */
+                                     .pre(path: "scripts/fix-tuist.sh", name: "Fix Tuist"),
+                                     .post(path: "scripts/lint.sh", name: "Swiftlint")
+                                 ],
+                                 dependencies: [
+                                     .target(name: "kDriveCore"),
+                                     .package(product: "FloatingPanel"),
+                                     .package(product: "Lottie"),
+                                     .package(product: "DropDown"),
+                                     .package(product: "HorizonCalendar")
+                                 ],
+                                 settings: .settings(base: shareExtensionSettings, debug: debugShareExtensionSettings)),
+                          Target(name: "kDriveActionExtension",
+                                 platform: .iOS,
+                                 product: .appExtension,
+                                 bundleId: "com.infomaniak.drive.ActionExtension",
+                                 deploymentTarget: .iOS(targetVersion: "13.0", devices: [.iphone, .ipad]),
+                                 infoPlist: .extendingDefault(with: [
+                                     "CFBundleVersion": "$(CURRENT_PROJECT_VERSION)",
+                                     "CFBundleShortVersionString": "$(MARKETING_VERSION)",
+                                     "CFBundleDisplayName": "Enregistrer dans kDrive",
+                                     "AppIdentifierPrefix": "$(AppIdentifierPrefix)",
+                                     "NSExtension": [
+                                         "NSExtensionMainStoryboard": "MainInterface",
+                                         "NSExtensionPointIdentifier": "com.apple.ui-services",
+                                         "NSExtensionAttributes": [
+                                             "NSExtensionActivationRule": "SUBQUERY (extensionItems, $extensionItem, SUBQUERY ($extensionItem.attachments, $attachment, (ANY $attachment.registeredTypeIdentifiers UTI-CONFORMS-TO \"public.data\")).@count == $extensionItem.attachments.@count ).@count > 0",
+                                             "NSExtensionServiceAllowsFinderPreviewItem": true,
+                                             "NSExtensionServiceAllowsTouchBarItem": true,
+                                             "NSExtensionServiceFinderPreviewIconName": "NSActionTemplate",
+                                             "NSExtensionServiceTouchBarBezelColorName": "TouchBarBezel",
+                                             "NSExtensionServiceTouchBarIconName": "NSActionTemplate"
+                                         ]
+                                     ]
+                                 ]),
+                                 sources: [
+                                     "kDriveActionExtension/**",
+                                     "kDrive/UI/Controller/FloatingPanelSelectOptionViewController.swift",
+                                     "kDrive/UI/Controller/Create File/FloatingPanelUtils.swift",
+                                     "kDrive/UI/Controller/Files/Categories/**",
+                                     "kDrive/UI/Controller/Files/Rights and Share/**",
+                                     "kDrive/UI/Controller/Files/Save File/**",
+                                     "kDrive/UI/Controller/Files/Search/**",
+                                     "kDrive/UI/Controller/Files/MultipleSelectionViewController.swift",
+                                     "kDrive/UI/Controller/Files/FileListViewController.swift",
+                                     "kDrive/UI/Controller/Files/FloatingPanelSortOptionTableViewController.swift",
+                                     "kDrive/UI/Controller/Floating Panel Information/**",
+                                     "kDrive/UI/Controller/NewFolder/**",
+                                     "kDrive/UI/Controller/Storyboard.swift",
+                                     "kDrive/UI/View/EmptyTableView/**",
+                                     "kDrive/UI/View/Header view/**",
+                                     "kDrive/UI/View/Generic/**",
+                                     "kDrive/UI/View/Files/Categories/**",
+                                     "kDrive/UI/View/Files/FileDetail/ShareLink/**",
+                                     "kDrive/UI/View/Files/SaveFile/**",
+                                     "kDrive/UI/View/Files/Search/**",
+                                     "kDrive/UI/View/Files/Upload/**",
+                                     "kDrive/UI/View/Files/FileCollectionViewCell.swift",
+                                     "kDrive/UI/View/Files/FileGridCollectionViewCell.swift",
+                                     "kDrive/UI/View/Files/SwipableCell.swift",
+                                     "kDrive/UI/View/Files/SwipableCollectionView.swift",
+                                     "kDrive/UI/View/Files/FloatingPanel/FloatingPanelSortOptionTableViewCell.swift",
+                                     "kDrive/UI/View/Files/FloatingPanel/FloatingPanelQuickActionCollectionViewCell.swift",
+                                     "kDrive/UI/View/Files/FloatingPanel/FloatingPanelTableViewCell.swift",
+                                     "kDrive/UI/View/Footer view/**",
+                                     "kDrive/UI/View/Menu/SwitchUser/**",
+                                     "kDrive/UI/View/Menu/MenuTableViewCell.swift",
+                                     "kDrive/UI/View/NewFolder/**",
+                                     "kDrive/Utils/**",
+                                     "Derived/Sources/Assets+KDrive.swift",
+                                     "Derived/Sources/Bundle+kDrive.swift",
+                                     "Derived/Sources/Strings+kDrive.swift"
+                                 ],
+                                 resources: [
+                                     "kDriveActionExtension/**/*.storyboard",
+                                     "kDrive/UI/Controller/Files/**/*.storyboard",
+                                     "kDrive/UI/Controller/Floating Panel Information/*.storyboard",
+                                     "kDrive/UI/Controller/NewFolder/*.storyboard",
+                                     "kDrive/UI/View/EmptyTableView/**/*.xib",
+                                     "kDrive/UI/View/Header view/**/*.xib",
+                                     "kDrive/UI/View/Generic/**/*.xib",
+                                     "kDrive/UI/View/Files/FileCollectionViewCell.xib",
+                                     "kDrive/UI/View/Files/FileGridCollectionViewCell.xib",
+                                     "kDrive/UI/View/Files/Categories/**/*.xib",
+                                     "kDrive/UI/View/Files/FileDetail/ShareLink/*.xib",
+                                     "kDrive/UI/View/Files/SaveFile/*.xib",
+                                     "kDrive/UI/View/Files/Search/*.xib",
+                                     "kDrive/UI/View/Files/Upload/*.xib",
+                                     "kDrive/UI/View/Files/FloatingPanel/FloatingPanelSortOptionTableViewCell.xib",
+                                     "kDrive/UI/View/Files/FloatingPanel/FloatingPanelQuickActionCollectionViewCell.xib",
+                                     "kDrive/UI/View/Files/FloatingPanel/FloatingPanelTableViewCell.xib",
+                                     "kDrive/UI/View/Footer view/*.xib",
+                                     "kDrive/UI/View/Menu/MenuTableViewCell.xib",
+                                     "kDrive/UI/View/Menu/SwitchUser/*.xib",
+                                     "kDrive/UI/View/NewFolder/*.xib",
+                                     "kDriveActionExtension/**/*.xcassets",
+                                     "kDrive/**/*.xcassets",
+                                     "kDriveActionExtension/**/*.strings",
+                                     "kDrive/**/Localizable.strings",
+                                     "kDrive/**/*.stringsdict",
+                                     "kDrive/**/*.json"
+                                 ],
+                                 entitlements: "kDriveActionExtension/ActionExtension.entitlements",
+                                 scripts: [
+                                     /* This prevents Tuist from generating automatic resources definition for this extension
+                                      as disabling it seems only possible at a project level (.disableSynthesizedResourceAccessors */
+                                     .pre(path: "scripts/fix-tuist.sh", name: "Fix Tuist"),
+                                     .post(path: "scripts/lint.sh", name: "Swiftlint")
+                                 ],
+                                 dependencies: [
+                                     .target(name: "kDriveCore"),
+                                     .package(product: "FloatingPanel"),
+                                     .package(product: "Lottie"),
+                                     .package(product: "DropDown"),
+                                     .package(product: "HorizonCalendar")
+                                 ],
+                                 settings: .settings(base: actionExtensionSettings, debug: debugActionExtensionSettings))
+                      ],
+                      fileHeaderTemplate: .file("file-header-template.txt"))
