+++ scripts/fix-tuist.sh
@@ -1,7 +1,7 @@
-echo "" > Derived/Sources/Bundle+kDriveActionExtension.swift
-echo "" > Derived/Sources/Assets+KDriveActionExtension.swift
-echo "" > Derived/Sources/Strings+KDriveActionExtension.swift
+echo "" > $SRCROOT/Derived/Sources/Bundle+kDriveActionExtension.swift
+echo "" > $SRCROOT/Derived/Sources/Assets+KDriveActionExtension.swift
+echo "" > $SRCROOT/Derived/Sources/Strings+KDriveActionExtension.swift
 
-echo "" > Derived/Sources/Bundle+kDriveShareExtension.swift
-echo "" > Derived/Sources/Assets+kDriveShareExtension.swift
-echo "" > Derived/Sources/Strings+kDriveShareExtension.swift
+echo "" > $SRCROOT/Derived/Sources/Bundle+kDriveShareExtension.swift
+echo "" > $SRCROOT/Derived/Sources/Assets+kDriveShareExtension.swift
+echo "" > $SRCROOT/Derived/Sources/Strings+kDriveShareExtension.swift
