+++ scripts/lint.sh
@@ -1,5 +1,5 @@
 if which swiftlint >/dev/null; then
-  swiftlint
+  swiftlint --config "$SRCROOT/.swiftlint.yml" --path "$SRCROOT/"
 else
   echo "warning: SwiftLint not installed, download from https://github.com/realm/SwiftLint"
 fi
